Silence some warnings about float to int casting

+++ TOMB4/game/laraflar.cpp
@@ -114,7 +114,7 @@ long DoFlareLight(PHD_VECTOR* pos, long flare_age)
 	}
 	else
 	{
-		falloff = flare_info.light_intensity - ((flare_age - (flare_info.flare_lifetime_in_ticks - 24)) >> 1) * ((float)flare_info.light_intensity / 16.0f);
+		falloff = flare_info.light_intensity - long(float(((flare_age - (flare_info.flare_lifetime_in_ticks - 24)) >> 1)) * ((float)flare_info.light_intensity / 16.0f));
 		r = (GetRandomControl() & 0x3F) + 64;
 		g = (GetRandomControl() & 0x3F) + 192;
 		b = GetRandomControl() & 0x1F;
@@ -138,7 +138,7 @@ long DoFlareLight(PHD_VECTOR* pos, long flare_age)
 	if (flare_info.has_sparks) {
 		if (flare_age < flare_info.flare_lifetime_in_ticks - 24) {
 			unsigned long flare_spark_rnd = GetRandomControl();
-			TriggerFlareSparks(pos->x, pos->y, pos->z, 0, flare_spark_rnd * -0.025, 0, flare_info.sparks_include_smoke);
+			TriggerFlareSparks(pos->x, pos->y, pos->z, 0, long((float(flare_spark_rnd) * -0.025f), 0, flare_info.sparks_include_smoke));
 		}
 	}
